# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Trader API
  description: |
    AI-powered stock recommendation service backend API
    
    This API provides comprehensive trading recommendation services, market data, 
    portfolio management, and user authentication features.
    
    ## Features
    - User authentication and profile management
    - Real-time stock market data integration via Finnhub API
    - AI-powered trading recommendations with multiple strategies
    - Portfolio tracking and performance analytics
    - Subscription-based access tiers
    - Advanced trading strategies (Momentum, Reversal, Swing Trading, etc.)
    
    ## Authentication
    This API uses JWT token-based authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```
    
    ## Rate Limiting
    - Authentication endpoints: 5 requests per minute
    - General API endpoints: 100 requests per minute
    - Market data endpoints: 200 requests per minute
    
    ## Subscription Tiers
    - **Basic**: 3 recommendations per day, Jesse Livermore strategy only
    - **Premium**: 50 recommendations per day, access to multiple strategies
    - **Professional**: Unlimited recommendations, all strategies, advanced analytics
    
  version: 1.0.0
  contact:
    name: Trader API Team
    email: support@trader-api.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api/v1
...
paths:
...
  /market/quote/{symbol}:
    get:
      tags:
        - Market Data
      summary: Get real-time stock quote
      description: Retrieve real-time price and market data for a stock symbol
      parameters:
        - name: symbol
          in: path
          required: true
          description: Stock symbol (e.g., AAPL, GOOGL)
          schema:
            type: string
            example: AAPL
      responses:
        '200':
          description: Quote data retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/StockQuote'
        '500':
          description: Failed to fetch market data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /market/candles/{symbol}:
    get:
      tags:
        - Market Data
      summary: Get candlestick chart data
      description: Retrieve historical candlestick data for technical analysis
      parameters:
        - name: symbol
          in: path
          required: true
          description: Stock symbol
          schema:
            type: string
        - name: resolution
          in: query
          description: Time resolution (1, 5, 15, 30, 60, D, W, M)
          schema:
            type: string
            default: D
            enum: ['1', '5', '15', '30', '60', 'D', 'W', 'M']
        - name: from
          in: query
          description: Start timestamp (Unix)
          schema:
            type: integer
        - name: to
          in: query
          description: End timestamp (Unix)
          schema:
            type: integer
      responses:
        '200':
          description: Candlestick data retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/CandlestickData'

  # Recommendations Endpoints
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error type
        message:
          type: string
          description: Human-readable error message
        code:
          type: string
          description: Error code for programmatic handling
...
    StockQuote:
      type: object
      properties:
        symbol:
          type: string
          example: AAPL
        current:
          type: number
          description: Current price
          example: 150.25
        change:
          type: number
          description: Price change from previous close
          example: 2.15
        percentChange:
          type: number
          description: Percentage change from previous close
          example: 1.45
        high:
          type: number
          description: Day's high price
          example: 152.30
        low:
          type: number
          description: Day's low price
          example: 148.90
        open:
          type: number
          description: Opening price
          example: 149.50
        previousClose:
          type: number
          description: Previous day's closing price
          example: 148.10
        timestamp:
          type: integer
          description: Last update timestamp (Unix)
    CandlestickData:
      type: object
      properties:
        o:
          type: array
          items:
            type: number
          description: Open prices
        h:
          type: array
          items:
            type: number
          description: High prices
        l:
          type: array
          items:
            type: number
          description: Low prices
        c:
          type: array
          items:
            type: number
          description: Close prices
        v:
          type: array
          items:
            type: number
          description: Volume data
        t:
          type: array
          items:
            type: integer
          description: Timestamp data
        s:
          type: string
          description: Status (ok/no_data)
...